pricemate/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
Appium-Python-Client==5.0.0
appwrite==11.0.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
attrs==24.2.0
Automat==24.8.1
azure-core==1.35.0
azure-storage-blob==12.26.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.35.70
botocore==1.35.70
Brotli==1.1.0
browser-cookie3==0.20.1
bs4==0.0.2
cachetools==4.2.4
celery==5.4.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cramjam==2.10.0
cryptography==43.0.3
cssselect==1.2.0
cssselect2==0.8.0
cycler==0.12.1
dateparser==1.2.0
DateTime==5.5
db-dtypes==1.4.1
deep-translator==1.11.4
defusedxml==0.7.1
deprecation==2.1.0
Django==5.2.4
django-tailwind==4.2.0
djangorestframework==3.16.0
dnspython==2.6.1
dropbox==12.0.2
dynamsoft_barcode_reader_bundle==10.4.2100
easygui==0.98.3
emoji==1.4.1
et-xmlfile==1.1.0
fake-useragent==1.5.1
fastapi==0.104.1
fastparquet==2024.11.0
filelock==3.16.1
fix-busted-json==0.0.18
Flask==3.1.1
fonttools==4.55.3
fpdf==1.7.2
frida==16.5.4
# frida-tools==13.3.0
frozenlist==1.7.0
fsspec==2025.5.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
# google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud==0.34.0
#google-cloud-bigquery==3.29.0
#google-cloud-core==2.4.1
#google-cloud-storage==2.18.2
#google-cloud-translate==2.0.1
#google-crc32c==1.6.0
#google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
# googletrans==4.0.0rc1
gotrue==2.12.3
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gspread==6.2.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib_metadata==8.7.0
incremental==24.7.2
isodate==0.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
kombu==5.4.2
langdetect==1.0.9
libretranslatepy==2.1.1
logging==0.4.9.6
lxml==5.4.0
lz4==4.4.3
MarkupSafe==3.0.2
matplotlib==3.10.0
motor==3.3.2
MouseInfo==0.1.3
msal==1.33.0
multidict==6.6.4
narwhals==1.43.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.4
parse==1.20.2
parsel==1.9.1
patool==4.0.1
pdfkit==1.0.0
pillow==11.0.0
# playwright==1.49.1
ply==3.11
postgrest==1.1.1
prompt_toolkit==3.0.48
propcache==0.3.2
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.0
psycopg2-binary==2.9.10
pushbullet.py==0.12.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
PyDrive==1.3.1
pydyf==0.11.0
pyee==12.0.0
pygeohash==2.1.0
PyGetWindow==0.0.9
Pygments==2.18.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pyjsparser==2.7.1
PyJWT==2.10.1
pymongo==4.6.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.5.0
pyperclip==1.9.0
pyphen==0.17.2
# pyppeteer==2.0.0
pyquery==2.0.1
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.6
python-telegram-bot==22.0
pytweening==1.2.0
pytz==2023.3
pywin32==308
PyYAML==6.0.2
pyzbar==0.1.9
# pyzxing==1.0.2
queuelib==1.7.0
random-user-agent==1.0.1
rarfile==4.2
#realtime==2.6.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.25.1
rsa==4.9
s3transfer==0.10.4
scapy==2.6.1
schedule==1.2.2
scipy==1.14.1
scraperapi-sdk==1.5.3
scrapingbee==2.0.1
Scrapy==2.11.2
selenium==4.27.1
selenium-stealth==1.0.6
service-identity==24.2.0
setuptools==75.3.0
shadowcopy==0.0.4
six==1.16.0
SkPy==0.11
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
starlette==0.27.0
stone==3.3.1
storage3==0.12.0
streamlit==1.46.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
svglib==1.5.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==2.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.1.3
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
translate==3.6.1
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
tzlocal==5.2
#undetected-chromedriver==3.5.5
Unidecode==1.3.8
uritemplate==4.1.1
uritools==5.0.0
urllib3==1.26.20
uvicorn==0.24.0
# vboxapi==1.0
vine==5.1.0
w3lib==2.2.1
watchdog==6.0.0
wcwidth==0.2.13
weasyprint==65.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
#websockets==10.4
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
xhtml2pdf==0.2.17
xlrd==2.0.1
XlsxWriter==3.2.0
xlwings==0.33.11
yarl==1.20.1
zipp==3.23.0
zope.interface==7.1.1
zopfli==0.2.3.post1
